Check for wrong template binding on StringRow
